Add XML serialization
We add the serialization part of the `xmlization` module so that we can
still write out XML on the micro-controller.

We can not include the de-serialization since there is no mature
XML parser for Micropython. The original XML serialization code writes
directly to a text stream, so this was easy to patch.