Extend rendering of formatted strings in patching (#14)
The patching string needs to consider more cases of the original source
code. Namely, we have to join strings together as micropython can not
deal with them (``"test" "me"`` needs to be represented as
``"testme"``). So far, we only handled variables (``ast.Name``) and
constants (``ast.Constant``), but recently we also have to handle a call
and an attribute access (``type(value).__name__``).

To that end, we extend the visitor for rendering formatted strings.